chore: Move all SCSS to be next to their component
