+++ src/components/common/container.module.scss
@@ -1,4 +1,4 @@
-@import '../../breakpoints.scss';
+@import '../../scss/breakpoints.scss';
 
 .container {
   max-width: 1000px;
