+++ src/components/common/detail-text.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import detailTextStyles from '../../scss/components/common/detail-text.module.scss'
+import detailTextStyles from './detail-text.module.scss'
 
 const DetailText = ({ children }) => (
   <div className={`detail-text ${detailTextStyles.detailText}`}>{children}</div>
