+++ src/components/common/detail-text.module.scss
@@ -1,4 +1,4 @@
-@import '../../colors.scss';
+@import '../../scss/colors.scss';
 
 .detail-text {
   font-size: 0.8rem;
