+++ src/components/common/disclosure.module.scss
