+++ src/components/common/infobox.module.scss
@@ -1,4 +1,4 @@
-@import '../../colors.scss';
+@import '../../scss/colors.scss';
 
 .infobox {
   display: inline-flex;
