+++ src/components/common/lead-paragraph.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import leadParagraphStyles from '../../scss/components/common/lead-paragraph.module.scss'
+import leadParagraphStyles from './lead-paragraph.module.scss'
 
 export default ({ children }) => (
   <p className={`lead-paragraph ${leadParagraphStyles.leadParagraph}`}>
