+++ src/components/common/press-list.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import pressListStyle from '../../scss/components/common/press-list.module.scss'
-import publicationFormattingStyle from '../../scss/components/common/publication-formatting.module.scss'
+import pressListStyle from './press-list.module.scss'
+import { PublicationTitle, PublicationSource } from './publication'
 import DetailText from './detail-text'
 
 export default ({ items }) => {
@@ -8,13 +8,11 @@ export default ({ items }) => {
     <ul className={`press-list ${pressListStyle.pressList}`}>
       {items.map(({ node }) => (
         <li key={`homepage-press-${node.id}`}>
-          <cite className={publicationFormattingStyle.title}>
+          <PublicationTitle>
             <a href={node.url}>{node.title}</a>
-          </cite>
+          </PublicationTitle>
           <DetailText>
-            <span className={publicationFormattingStyle.source}>
-              {node.publication}
-            </span>
+            <PublicationSource>{node.publication}</PublicationSource>
             <span className={pressListStyle.dotSeparator}>•</span>
             {node.publishDate}
           </DetailText>
