+++ src/components/common/press-list.module.scss
