+++ src/components/common/publication.module.scss
