+++ src/components/common/table.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import DetailText from './detail-text'
-import tableStyle from '../../scss/components/common/table.module.scss'
+import tableStyle from './table.module.scss'
 
 const Table = ({ children, tableLabel }) => (
   <div>
