+++ src/components/layout/header.module.scss
@@ -1,5 +1,5 @@
-@import '../colors.scss';
-@import '../breakpoints.scss';
+@import '../../scss/colors.scss';
+@import '../../scss/breakpoints.scss';
 
 .circles {
   background: #5667ad;
