+++ src/components/layout/hero.module.scss
@@ -1,5 +1,5 @@
-@import '../../colors.scss';
-@import '../../breakpoints.scss';
+@import '../../scss/colors.scss';
+@import '../../scss/breakpoints.scss';
 
 .hero {
   @media only screen and (max-width: $viewport-ms) {
