+++ src/pages/data/index.module.scss
@@ -1,5 +1,5 @@
-@import '../../../colors.scss';
-@import '../../../breakpoints.scss';
+@import '../../scss/colors.scss';
+@import '../../scss/breakpoints.scss';
 
 .state-nav {
   position: relative;
@@ -25,33 +25,5 @@
   .state-nav-inner {
     border-bottom: 2px solid $color-grey-6;
     width: 100%;
-  }
-}
-
-.state-nav-combobox {
-  label {
-    font-weight: bold;
-    display: inline-block;
-  }
-
-  [data-reach-combobox] {
-    margin: 0.3rem 0;
-    @media (min-width: $viewport-ms) {
-      display: inline-block;
-      margin-left: 1rem;
-    }
-  }
-  input {
-    padding: 0.3rem;
-    border: 1px solid $color-grey-4;
-  }
-}
-
-.popover {
-  [data-user-value] {
-    font-weight: bold;
-  }
-  [data-suggested-value] {
-    font-weight: normal;
   }
 }
