--- src/scss/components/development-warning.module.scss
@@ -1,8 +1,0 @@
-.development-warning {
-  color: #fff;
-  background: #bd2c2c;
-  padding: 1rem;
-  a {
-    color: #fff;
-  }
-}
